requirements.txt
django>=5.2,<6.0
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-decouple==3.8
yfinance>=0.2.54
feedparser==6.0.11
requests==2.32.3
groq==0.11.0
gunicorn==22.0.0
whitenoise==6.7.0
pandas>=2.0.0
numpy>=1.24.0
dj-database-url>=2.0.0
psycopg2-binary>=2.9.0